Move FullScreenAlarmScreen related component files to new component directory
